# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portfolio Manager API",
    "description": "Phase 1 backend. Not financial advice.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/portfolios": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get Portfolios",
        "operationId": "get_portfolios_api_portfolios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioOut"
                  },
                  "type": "array",
                  "title": "Response Get Portfolios Api Portfolios Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio",
        "operationId": "create_portfolio_api_portfolios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/portfolios/{portfolio_id}": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Get One",
        "operationId": "get_one_api_portfolios__portfolio_id__get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "portfolios"
        ],
        "summary": "Patch Portfolio",
        "operationId": "patch_portfolio_api_portfolios__portfolio_id__patch",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Delete Portfolio",
        "operationId": "delete_portfolio_api_portfolios__portfolio_id__delete",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/summary": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Summary",
        "operationId": "summary_api_portfolios__portfolio_id__summary_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/holdings": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Holdings",
        "operationId": "holdings_api_portfolios__portfolio_id__holdings_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HoldingOut"
                  },
                  "title": "Response Holdings Api Portfolios  Portfolio Id  Holdings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Add Holding",
        "operationId": "add_holding_api_portfolios__portfolio_id__holdings_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/xray": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Xray",
        "operationId": "xray_api_portfolios__portfolio_id__xray_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XrayOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/heatmap": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "Heatmap",
        "operationId": "heatmap_api_portfolios__portfolio_id__heatmap_get",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeatmapItem"
                  },
                  "title": "Response Heatmap Api Portfolios  Portfolio Id  Heatmap Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/holdings/{holding_id}": {
      "patch": {
        "tags": [
          "portfolios"
        ],
        "summary": "Patch Holding",
        "operationId": "patch_holding_api_portfolios__portfolio_id__holdings__holding_id__patch",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "holding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Holding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldingPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolios"
        ],
        "summary": "Delete Holding",
        "operationId": "delete_holding_api_portfolios__portfolio_id__holdings__holding_id__delete",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "holding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Holding Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name_en": {
            "type": "string",
            "title": "Name En"
          },
          "name_he": {
            "type": "string",
            "title": "Name He"
          },
          "asset_type": {
            "type": "string",
            "enum": [
              "stock",
              "etf",
              "crypto",
              "fund",
              "bond",
              "cash"
            ],
            "title": "Asset Type"
          },
          "market": {
            "type": "string",
            "enum": [
              "US",
              "TASE",
              "CRYPTO"
            ],
            "title": "Market"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "day_change_pct": {
            "type": "number",
            "title": "Day Change Pct"
          },
          "value_ils": {
            "type": "number",
            "title": "Value Ils"
          },
          "pnl": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Pnl"
              },
              {
                "type": "null"
              }
            ]
          },
          "weight_pct": {
            "type": "number",
            "title": "Weight Pct"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "1w",
                  "1m",
                  "3m",
                  "6m",
                  "1y"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "stop_tp_status": {
            "type": "string",
            "enum": [
              "missing",
              "needs_horizon"
            ],
            "title": "Stop Tp Status"
          },
          "score_card": {
            "$ref": "#/components/schemas/ScoreCardMini"
          },
          "price_stale": {
            "type": "boolean",
            "title": "Price Stale",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "name_en",
          "name_he",
          "asset_type",
          "market",
          "quantity",
          "price",
          "currency",
          "day_change_pct",
          "value_ils",
          "weight_pct",
          "stop_tp_status",
          "score_card"
        ],
        "title": "HoldingOut"
      },
...
      "PortfolioCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "base_currency": {
            "type": "string",
            "enum": [
              "ILS",
              "USD"
            ],
            "title": "Base Currency",
            "default": "ILS"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PortfolioCreate"
      },
      "PortfolioOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "base_currency": {
            "type": "string",
            "enum": [
              "ILS",
              "USD"
            ],
            "title": "Base Currency"
          },
          "risk_filter": {
            "$ref": "#/components/schemas/RiskFilterOut"
          },
          "tracking_started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Started At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "base_currency",
          "risk_filter",
          "created_at"
        ],
        "title": "PortfolioOut"
      },
      "PortfolioPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "base_currency": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ILS",
                  "USD"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Currency"
          },
          "risk_filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskFilterIn"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PortfolioPatch"
      },
...
      "SummaryOut": {
        "properties": {
          "value": {
            "$ref": "#/components/schemas/Money"
          },
          "day_pnl": {
            "$ref": "#/components/schemas/Pnl"
          },
          "week_pnl": {
            "$ref": "#/components/schemas/Pnl"
          },
          "month_pnl": {
            "$ref": "#/components/schemas/Pnl"
          },
          "since_start_pnl": {
            "$ref": "#/components/schemas/Pnl"
          },
          "since_start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since Start Date"
          },
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "fx_stale": {
            "type": "boolean",
            "title": "Fx Stale"
          },
          "weekly_bars": {
            "items": {
              "$ref": "#/components/schemas/WeeklyBar"
            },
            "type": "array",
            "title": "Weekly Bars"
          },
          "monthly_bars": {
            "items": {
              "$ref": "#/components/schemas/MonthlyBar"
            },
            "type": "array",
            "title": "Monthly Bars"
          },
          "since_start_series": {
            "items": {
              "$ref": "#/components/schemas/SeriesPoint"
            },
            "type": "array",
            "title": "Since Start Series"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "markets": {
            "$ref": "#/components/schemas/Markets"
          }
        },
        "type": "object",
        "required": [
          "value",
          "day_pnl",
          "week_pnl",
          "month_pnl",
          "since_start_pnl",
          "week_start",
          "fx_stale",
          "weekly_bars",
          "monthly_bars",
          "since_start_series",
          "as_of",
          "markets"
        ],
        "title": "SummaryOut"
      },
...
    }
  }
}